feat: optimize the authOptions
