+++ app/api/auth/[...nextauth]/route.ts
@@ -1,51 +1,5 @@
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import { PrismaClient } from "@prisma/client";
-import { AuthOptions } from "next-auth";
+import { authOptions } from "@/lib/auth";
 import NextAuth from "next-auth/next";
-import GithubProvider from "next-auth/providers/github";
-import GoogleProvider from "next-auth/providers/google";
-
-const prisma = new PrismaClient();
-
-export const authOptions: AuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID as string,
-      clientSecret: process.env.GITHUB_CLIENT_SECRET as string,
-    }),
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-  ],
-  session: {
-    strategy: "jwt",
-  },
-  pages: {
-    signIn: "/sign-in",
-  },
-  callbacks: {
-    async session({ session, token }) {
-      if (token.sub && session.user) {
-        session.user.id = token.sub;
-      }
-      return session;
-    },
-    async redirect({ url, baseUrl }) {
-      // 登录成功后，始终重定向到 / 路由
-      if (url.startsWith(baseUrl)) {
-        return `${baseUrl}/`;
-      }
-      // 处理来自外部链接的重定向
-      else if (url.startsWith('/')) {
-        return `${baseUrl}${url}`;
-      }
-      return url;
-    }
-  },
-  secret: process.env.NEXTAUTH_SECRET,
-};
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
